pyproject.toml
[project]
name = "ps-of-py"
version = "0.2.0"
description = "This is a  Python  for Photoshop."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "photoshop-python-api>=0.24.1",
    "xlwings>=0.33.15",
]

[dependency-groups]
dev = [
    "coverage>=7.10.7",
    "pyinstaller>=6.16.0",
]
